Bring GoogleMaps logic in to HomePage and delete common component.
