--- mobile/src/app/common/google-maps/google-maps.component.scss
@@ -1,3 +1,0 @@
-#map {
-  height: 100%;
-}
