+++ mobile/src/app/pages/home/home.module.ts
@@ -7,7 +7,6 @@ import { HomePage } from './home.page';
 import { HomePageRoutingModule } from './home-routing.module';
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
-import { GoogleMapsModule } from '../../common/google-maps/google-maps.module';
 
 @NgModule({
   imports: [
@@ -15,7 +14,6 @@ import { GoogleMapsModule } from '../../common/google-maps/google-maps.module';
     FormsModule,
     IonicModule,
     HomePageRoutingModule,
-    GoogleMapsModule,
     CommonModule,
     // import HttpClientModule after BrowserModule.
     HttpClientModule
