+++ mobile/src/app/pages/home/home.page.html
@@ -7,8 +7,26 @@
 </ion-header>
 
 <ion-content [fullscreen]="true">
-  <app-google-maps
-    [data]="data$ | async"
-    (refresh$)="refresh()"
-  ></app-google-maps>
+  <div #map id="map" style="height: 100%;"></div>
+  <div *ngIf="selectionRectangle">
+    <ion-fab vertical="bottom" horizontal="center">
+      <ion-button (click)="onAdd()"
+        >Place '{{ currentMenuKey }}' Pin.</ion-button
+      >
+      <ion-button color="danger" (click)="onCancel()">X</ion-button>
+    </ion-fab>
+  </div>
+  <ion-fab vertical="bottom" horizontal="end">
+    <ion-fab-button>
+      <ion-icon name="add-outline"></ion-icon>
+    </ion-fab-button>
+    <ion-fab-list side="top">
+      <ion-fab-button (click)="paramsChanged('water')">
+        <ion-icon name="water-outline"></ion-icon>
+      </ion-fab-button>
+      <ion-fab-button (click)="paramsChanged('first aid')">
+        <ion-icon name="medkit-outline"></ion-icon>
+      </ion-fab-button>
+    </ion-fab-list>
+  </ion-fab>
 </ion-content>
